// src/app/modules/clients/model/EducationOrQualificationModel.ts
import { MlcModel } from './common/mlc.model';

export class EducationOrQualification implements MlcModel {
  name: string;
  institution: string;
  locationoraddress: string;
  datestarted: string;
  datefinishedorexpectedfinish: string;
  studentid: string;
  studentaccountorlogins: string;
  accommodationname: string;
  accommodationcontact: string;
  accommodationaddreess: string;
  transportdetails: string;
  parkingdetails: string;
  results: string;
  highlights: string;
  notes: string;
}